requirements.txt
Babel==2.8.0
cycler==0.10.0
fpdf==1.7.2
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.4
pandas==1.0.3
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
tkcalendar==1.6.1
ttkthemes==3.0.0
